fix bugs

+++ app/src/main/java/com/codeest/geeknews/util/SystemUtil.java
@@ -8,6 +8,8 @@ import android.graphics.Bitmap;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.net.Uri;
+import android.provider.MediaStore;
+import android.view.View;
 
 import com.codeest.geeknews.app.App;
 import com.codeest.geeknews.app.Constants;
@@ -24,10 +26,6 @@ public class SystemUtil {
 
     /**
      * 检查WIFI是否连接
-     *
-     * @author Ysjian
-     * @date 2014-5-9
-     * @return 如果连接了返回true，否则返回false
      */
     public static boolean isWifiConnected() {
         ConnectivityManager connectivityManager = (ConnectivityManager) App.getInstance().getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
@@ -37,10 +35,6 @@ public class SystemUtil {
     }
     /**
      * 检查手机网络(4G/3G/2G)是否连接
-     *
-     * @author Ysjian
-     * @date 2014-5-9
-     * @return 如果连接了返回true，否则返回false
      */
     public static boolean isMobileNetworkConnected() {
         ConnectivityManager connectivityManager = (ConnectivityManager) App.getInstance().getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
@@ -50,10 +44,6 @@ public class SystemUtil {
     }
     /**
      * 检查是否有可用网络
-     *
-     * @author Ysjian
-     * @date 2014-5-9
-     * @return 存在WIFI和手机数据任意可用网络返回true，否则返回false
      */
     public static boolean isNetworkConnected() {
         ConnectivityManager connectivityManager = (ConnectivityManager) App.getInstance().getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
@@ -78,29 +68,36 @@ public class SystemUtil {
      * @param url
      * @param bitmap
      */
-    public static Uri saveBitmapToFile(Context context, String url, Bitmap bitmap){
+    public static Uri saveBitmapToFile(Context context, String url, Bitmap bitmap,View container, boolean isShare){
         String fileName = url.substring(url.lastIndexOf("/"),url.lastIndexOf(".")) + ".png";
         File fileDir = new File(Constants.PATH_DATA);
         if (!fileDir.exists()){
             fileDir.mkdir();
         }
         File imageFile = new File(fileDir,fileName);
+        Uri uri = Uri.fromFile(imageFile);
+        if (isShare && imageFile.exists()) {
+            return uri;
+        }
         try {
             FileOutputStream fos = new FileOutputStream(imageFile);
-            boolean isCompress = bitmap.compress(Bitmap.CompressFormat.PNG, 256, fos);
+            boolean isCompress = bitmap.compress(Bitmap.CompressFormat.PNG, 90, fos);
             if (isCompress) {
-                ToastUtil.shortShow("图片保存成功");
+                SnackbarUtil.showShort(container,"保存妹纸成功n(*≧▽≦*)n");
             } else {
-                ToastUtil.shortShow("图片保存失败");
+                SnackbarUtil.showShort(container,"保存妹纸失败ヽ(≧Д≦)ノ");
             }
             fos.flush();
             fos.close();
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
+            SnackbarUtil.showShort(container,"保存妹纸失败ヽ(≧Д≦)ノ");
         }
-        Uri uri = Uri.fromFile(imageFile);
+        try {
+            MediaStore.Images.Media.insertImage(context.getContentResolver(), imageFile.getAbsolutePath(), fileName, null);
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
         context.sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE,uri));
         return uri;
     }
